pyproject.toml
[build-system]
requires = ["setuptools>=71.1.0","setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "crcutil"
version = "1.7.0"
requires-python = ">=3.12.10"
authors = [
  {name = "Carlos Florez", email = "carlos@florez.co.uk"}
]
maintainers = [
  {name = "Carlos Florez", email = "carlos@florez.co.uk"}
]
description = ""
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "alive-progress==3.1.5",
  "pyyaml==6.0.1",
  "toml==0.10.2",
  "pywin32==306;sys_platform == 'win32'",
  "jsonschema==4.23.0",
  "colorlog==6.4.0",
  "pynput==1.7.7",
  "python-xlib==0.33",
  "evdev==1.9.2",
]

[project.scripts]
crcutil = "crcutil.__main__:main"

[project.urls]
Repository = "https://github.com/florez-carlos/crcutil"
